organise pages and header into separate folder / component
