+++ src/pages/Leaders.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Grid from '@material-ui/core/Grid';
-import BasicCard from './components/BasicCard';
+import BasicCard from '../components/BasicCard';
 
 const Leaders = () => (
   <main>
